+++ dotnet-docs-viewer/Dto/Common/FileInfoDto.cs
@@ -7,7 +7,7 @@ namespace WebApp.Dto
     [JsonObject]
     public abstract class FileInfoBaseDto : BaseDto
     {
-        public virtual long FileCode { get; protected set; }
+        public virtual int FileCode { get; protected set; }
         public virtual string FileName { get; protected set; } = null!;
         public virtual string FilePath { get; protected set; } = null!;
         public virtual string FileType { get; protected set; } = null!;
